Fixed TWEE.IO
It is now upper cased, because it become mark, TWEE - means The Web Elegant Engine. 
Also twee translates as elegant too.
So it is required to be upper cased. 

Also it became full-stack, so I added one more link to appropriate place according it's number of stars on GitHub.

Thank you!